backend/parseXML/requirements.txt
arrow==1.3.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
DateTime==5.2
dnspython==2.4.2
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
zope.interface==6.1
